requirements.txt
amqp==5.0.5
asgiref==3.3.1
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Django==3.1.6
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.4.0
django-restframework==0.0.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
ftps==0.1.0
gunicorn==20.0.4
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
Markdown==3.3.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.14
psycopg2==2.8.6
ptyprocess==0.7.0
pycurl==7.43.0.6
Pygments==2.7.4
PyJWT==2.0.1
pyTelegramBotAPI==3.7.6
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
